+++ src/utils/hooks.ts
@@ -6,6 +6,9 @@ import { Range } from '../types';
  * Key Press
  */
 export const useKeyPress = (key: string, onKeyPressed: (pressed: boolean) => void) => {
+  /**
+   * Key down
+   */
   const keydownListener = useCallback(
     (e: any) => {
       if (e.key === 'Shift') {
@@ -15,6 +18,9 @@ export const useKeyPress = (key: string, onKeyPressed: (pressed: boolean) => voi
     [onKeyPressed]
   );
 
+  /**
+   * Key up
+   */
   const keyupListener = useCallback(
     (e: any) => {
       if (e.key === 'Shift') {
@@ -24,11 +30,17 @@ export const useKeyPress = (key: string, onKeyPressed: (pressed: boolean) => voi
     [onKeyPressed]
   );
 
+  /**
+   * Event for Key down
+   */
   useEffect(() => {
     window.addEventListener('keydown', keydownListener, true);
     return () => window.removeEventListener('keydown', keydownListener, true);
   }, [keydownListener]);
 
+  /**
+   * Event for Key up
+   */
   useEffect(() => {
     window.addEventListener('keyup', keyupListener, true);
     return () => window.removeEventListener('keyup', keyupListener, true);